pyproject.toml
[project]
name = "penpal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.7.0",
]
